refactor: rename constants folder to config
