+++ src/Maintenance.app.js
@@ -10,7 +10,7 @@ import { Overview } from './pages/Overview'
 import { ProtectedRoute } from './modules/Navigation/ProtectedRoute'
 import { RedirectToOld } from './pages/RedirectToOld'
 import { configureStore, history } from './redux/configureStore'
-import { dataElementSections } from './constants/sections'
+import { dataElementSections } from './config/sections'
 import styles from './Maintenance.module.css'
 
 const store = configureStore()
