+++ src/pages/All.js
@@ -3,8 +3,8 @@ import React from 'react'
 import { CardMenu } from '../modules/CardMenu/CardMenu'
 import { Container } from '../modules/layout/Container'
 import { Content } from '../modules/layout/Content'
-import { groupOrder } from '../constants/groupOrder'
-import { sectionOrder } from '../constants/sectionOrder'
+import { groupOrder } from '../config/groupOrder'
+import { sectionOrder } from '../config/sectionOrder'
 
 export const All = () => (
     <Container>
