+++ src/pages/Overview.js
@@ -6,7 +6,7 @@ import { SideBar as SideBarLayout } from '../modules/layout/SideBar'
 import { MainContent } from '../modules/ContentWrappers/MainContent'
 import { Sidebar } from '../modules/ContentWrappers/Sidebar'
 import { CardMenu } from '../modules/CardMenu/CardMenu'
-import { sectionOrder } from '../constants/sectionOrder'
+import { sectionOrder } from '../config/sectionOrder'
 
 export const Overview = ({ name, match }) => {
     const { group } = match.params
